fix: Legacy Asset Loading
Creates a reusable component that loads both the logacy JS and CSS
assets for modern pages.
